refactor: move guest-cleanup to users
