+++ backend/src/app.module.ts
@@ -11,7 +11,7 @@ import { TfaModule } from './auth/tfa/tfa.module';
 import { AppConfigModule } from './config/app-config.module';
 import { TypeOrmConfigModule } from './config/typeorm-config.module';
 import { FriendRequestModule } from './friend_requests/friend_request.module';
-import { GuestCleanupService } from './guest-cleanup/guest-cleanup.service';
+import { GuestCleanupService } from './users/guest-cleanup/guest-cleanup.service';
 import { ScheduleModule } from '@nestjs/schedule';
 import { StatusModule } from './users/status/status.module';
